+++ src/assets/svg/Icons/DiscourseIcon.tsx
