+++ src/assets/svg/Icons/SlackIcon.tsx
